Feat: Mejoras en el código, correción de errores
