+++ src/app/app.component.html
@@ -11,7 +11,6 @@
   <li class="nav-item">
     <a class="nav-link" routerLink="/profile">Perfil</a>
   </li>
-  <app-navigation-menu></app-navigation-menu>
   <router-outlet></router-outlet>
 </ul>
