--- src/app/home/home.component.html
@@ -1,3 +1,0 @@
-<h1>Página de Inicio</h1>
-<p>Bienvenido/a a nuestra aplicación de planificación de comidas saludables.</p>
-
