--- src/app/meal-plans/meal-plan.module.ts
@@ -1,11 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-
-@NgModule({
-  imports: [
-  CommonModule,
-  // Other modules you may have
-  ],
-})
-export class MealPlanModule { }
-
