--- src/app/meal-plans/meal-plan.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { MealPlanService } from './meal-plan.service';
-
-describe('MealPlamService', () => {
-  let service: MealPlanService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(MealPlanService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
