--- src/app/navigation-menu/navigation-menu.component.html
@@ -1,6 +1,0 @@
-<nav>
-  <ul>
-    <li><a routerLink="/">Inicio</a></li>
-    <li><a routerLink="/recomendaciones">Recomendaciones de Planes de Comidas</a></li>
-  </ul>
-</nav>
