--- src/app/navigation-menu/navigation-menu.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { NavigationMenuComponent } from './navigation-menu.component';
-
-describe('NavigationMenuComponent', () => {
-  let component: NavigationMenuComponent;
-  let fixture: ComponentFixture<NavigationMenuComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [NavigationMenuComponent]
-    });
-    fixture = TestBed.createComponent(NavigationMenuComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
