--- src/app/navigation-menu/navigation-menu.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-navigation-menu',
-  templateUrl: './navigation-menu.component.html',
-  styleUrls: ['./navigation-menu.component.css']
-})
-export class NavigationMenuComponent {
-
-}
