--- src/app/profile/profile.component.css
@@ -1,34 +1,0 @@
-/* Estilos para el formulario */
-.form-container {
-    max-width: 400px;
-    margin: 0 auto;
-  }
-  
-  .form-field {
-    margin-bottom: 20px;
-  }
-  
-  /* Estilos para los campos de edición */
-  .mat-input-element {
-    border-radius: 5px;
-    padding: 10px;
-    border: 1px solid #ccc;
-  }
-  
-  .mat-select {
-    width: 100%;
-  }
-  
-  /* Estilos para el botón de guardar cambios */
-  .save-button {
-    background-color: #007bff;
-    color: #fff;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .save-button:hover {
-    background-color: #0056b3;
-  }
