--- src/app/profile/profile.component.html
@@ -1,38 +1,0 @@
-<div class="container">
-  <header>
-    <h1>Actualización de perfil</h1>
-  </header>
-  <main>yy
-    <form (ngSubmit)="updateProfile()">
-      <div class="form-group">
-        <mat-form-field>
-          <input matInput placeholder="Nombre" formControlName="nombre" required>
-        </mat-form-field>
-      </div>
-      <div class="form-group">
-        <mat-form-field>
-          <input matInput placeholder="Dirección" formControlName="direccion" required>
-        </mat-form-field>
-      </div>
-      <div class="form-group">
-        <mat-form-field>
-          <input matInput placeholder="Teléfono" formControlName="telefono" required>
-        </mat-form-field>
-      </div>
-      <div class="form-group">
-        <mat-form-field>
-          <input matInput type="email" placeholder="Correo electrónico" formControlName="correo" required>
-        </mat-form-field>
-      </div>
-      <button type="submit" class="btn btn-primary">Guardar cambios</button>
-    </form>
-  </main>
-  <footer>
-    <p>Derechos de autor © 2023. Todos los derechos reservados.</p>
-    <nav>
-      <a href="#">Términos y condiciones</a> |
-      <a href="#">Política de privacidad</a> |
-      <a href="#">Contáctanos</a>
-    </nav>
-  </footer>
-</div>
