--- src/app/profile/profile.model.ts
@@ -1,7 +1,0 @@
-export interface Profile {
-    nombre: string;
-    direccion: string;
-    telefono: string;
-    correo: string;
-  }
-
