--- src/app/profile/profile.module.ts
@@ -1,26 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatInputModule } from '@angular/material/input';
-import { ProfileComponent } from './profile.component';
-import { MatFormFieldModule } from '@angular/material/form-field';
-
-
-@NgModule({
-  declarations: [
-    ProfileComponent
-  ],
-  imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatInputModule,
-    MatFormFieldModule,
-  ],
-  exports: [
-    ProfileComponent
-  ]
-})
-export class ProfileModule { }
-
-
