--- src/app/profile/profile.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ProfileService } from './profile.service';
-
-describe('ProfileService', () => {
-  let service: ProfileService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ProfileService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
