--- src/app/profile/profile.service.ts
@@ -1,21 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
-import { Profile } from './profile.model';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class ProfileService {
-  private apiUrl = 'http://localhost:8080';
-
-  constructor(private http: HttpClient) { }
-
-  getProfile(): Observable<Profile> {
-    return this.http.get<Profile>(`${this.apiUrl}/profile`);
-  }
-
-  updateProfile(profileData: Profile): Observable<any> {
-    return this.http.put<any>(`${this.apiUrl}/profile`, profileData);
-  }
-}
